functions/requirements.txt
firebase_functions~=0.1.0
firebase_admin
flask
flask[async]
python-dotenv
flask-cors
PyPDF2
docx2txt
pdfminer.six
google-cloud-storage
yake
bz2file
langchain-openai 
langchain-pinecone  
pinecone-client 
openai 
langchain-google-firestore
langchain_google_firestore 
langchain-pinecone 
langchain-openai 
langchain
langchain-community 
langchainhub 
langchain-core
google-auth
google-cloud-firestore
sentence-transformers